// types/supabase-internal-types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: '13.0.5'
  }
  internal: {
    Tables: {
// ...
      ws_notifications_ios: {
        Row: {
          announcement_last_delivery_at: string | null
          announcement_notifications: boolean | null
          created_at: string
          daily_chapter_last_delivery_at: string | null
          daily_chapter_notifications: boolean | null
          daily_verse_last_delivery_at: string | null
          daily_verse_notifications: boolean | null
          device_token: string
          is_sandbox: boolean
          last_delivery_at: string | null
          prayer_times_last_delivery_at: string | null
          prayer_times_notifications: Json
          updated_at: string | null
        }
        Insert: {
          announcement_last_delivery_at?: string | null
          announcement_notifications?: boolean | null
          created_at?: string
          daily_chapter_last_delivery_at?: string | null
          daily_chapter_notifications?: boolean | null
          daily_verse_last_delivery_at?: string | null
          daily_verse_notifications?: boolean | null
          device_token: string
          is_sandbox?: boolean
          last_delivery_at?: string | null
          prayer_times_last_delivery_at?: string | null
          prayer_times_notifications?: Json
          updated_at?: string | null
        }
        Update: {
          announcement_last_delivery_at?: string | null
          announcement_notifications?: boolean | null
          created_at?: string
          daily_chapter_last_delivery_at?: string | null
          daily_chapter_notifications?: boolean | null
          daily_verse_last_delivery_at?: string | null
          daily_verse_notifications?: boolean | null
          device_token?: string
          is_sandbox?: boolean
          last_delivery_at?: string | null
          prayer_times_last_delivery_at?: string | null
          prayer_times_notifications?: Json
          updated_at?: string | null
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}